First commit of the Treatment Placement notebook
Workflow for the determination of optimum treatment placement for
management options across the Dinkey Creek watershed. First commit,
looking at ways to begin to define management units within the
watershed.

Updated severityGen to severityPGen, added the functionality of
calculating the probability of severe fire (4 and 5) within the same
raster stack that is used to calculate mean and variance of burn
severity.