Added year 100 AGB plots and tests
Had to tweak genWeightedCentOutput to return a dataframe that wasn't
averaged or summed across replicates. I needed to look at year 100
biomass from century, and compare across treatments within fire weather
differences. Basically added some lines to that function that allowed
this capability., and made a couple boxplots to visualize the
distribution of end of simulation AGB values. Also added in a Tukey post
hoc test to assess the significance of any differences.