Added error handling to rH modeler
modeling rH using temp and precip is messy work, and sometimes there is
a lack of OLS model convergence -- this happens, and I'm building in
steps to gap fill months etc. Currently however, it is no longer a
barrier to progress and I handle errors by filling the fit statistic
fields with -9999 for processing later.